+++ database/migrations/2026_02_12_000003_create_budgets_table.php
@@ -11,7 +11,7 @@ return new class extends Migration
         Schema::create('budgets', function (Blueprint $table) {
             $table->id();
             $table->foreignId('category_id')->constrained()->cascadeOnDelete();
-            $table->integer('amount'); // stored in centavos
+            $table->decimal('amount', 10, 2);
             $table->integer('month');
             $table->integer('year');
             $table->timestamps();
